Fix some mistakes in the capnp files, add a type parameter to AppHooks.
I think I was still fighting with the build system when I first drafted
this, so didn't catch things like missing ordinals, syntax errors, etc.
This fixes those.

The BridgeObjectId defined in sandstorm-http-bridge.capnp was moved into
the scope of AppHooks, to avoid a name collision with the one in
sandstorm-http-bridge-internal.capnp.

Finally, AppHooks is given a type parameter, which is then used as the
argument to the `application`. Variant. This is a bit more type safe
than what we had before.