Freshen docs build to work with mkdocs 1.0
The `theme_dir` option changed to `theme.custom_dir` [1], and the extension
called `inline_graphviz` is apparently now called `mdx_inline_graphviz`[2] [3].

(Additionally, the `pages` option was renamed to `nav` which is not yet
breaking, but they threaten to break it in the future.)

The modifications to the readthedocs theme (from circa 0.17) have been ported
forwawrd to the theme as-of 1.0.4. This required renaming README.md to
readme.txt to avoid fouling the landing page autodetection.

Also, pinned the versions in the pip install commands, to reduce future surprise
breakage of this type.

1: https://www.mkdocs.org/about/release-notes/#theme_dir-configuration-option-fully-deprecated

2: https://github.com/sprin/markdown-inline-graphviz/issues/6

3: https://github.com/sprin/markdown-inline-graphviz/issues/9