Go back to meteor-spk 0.3.2
For some reason with the upgraded version, we end up pulling in a new
version of node but not of the meteor dev bundle, so node fibers bails
out because of an ABI mismatch. We should update at somepoint, but this
is starting to feel like too much yak shaving, and meteor-spk itself
could use an upgrade anyway.