meteor-testapp: set expectAppHooks = true, and fix a segfault in the bridge.
I'm not really sure what I was thinking when I wrote the first pass of
this; it was blatantly using the appHooks promise after it had gone out
of scope.

This still segfaults, but later. I'm 95% sure this is because we're
never calling fork() on the promise, so the second time we try to
resolve it it is no longer valid.