Fixes to the bridge's membrane.
We really want everything the app provides to be inside the membrane,
and everything else (including the bridge cap) to be outside of it.
This patch corrects that.

Previously, stuff returned by AppHooks.restore() would not have been
wrapped. I don't think this currently matters since we never call save()
on stuff returned from restore(), but down the line it could cause
problems.