Change obviously wrong implementation of outboundCall.
I'm not sure what I was thinking when writing the first version. Some
version of "shouldn't just return null because of the comments
about asymmetry in membrane.h," but double-wrapping the capability
doesn't actually help with that.

Per the comment, after thinking about how to have this behave
"correctly" for a bit, I came to the conclusion that it wasn't worth
the trouble, and wasn't necessarily the desired behavior anyway.