Lift blackrock-payments Meteor package into imports
To get the remaining packages into imports, we need to hoist them up in the
reverse order they are loaded.  `blackrock-payments` is a leaf in that
dependency tree; let's lift that one first.

We move Blackrock's Stripe dependency into sandstorm's top-level package.json.
We should be able to swap out Npm.require() for plain old imports.

The two asset files are placed in private/, per the Meteor Guide.