Lift sandstorm-db out from packages into imports/
This appeared to have a dependency on a particular version of the
content-type package from npm.  We can sort that out separately.

Additionally, this has a Meteor package dependency on
fongandrew:find-and-modify.  I pinned the version because I vaguely recall this
particular library breaking in the past when we tried to update it and don't
want to deal with that right now either.