Update fourseven:scss to 4.12.0.
Version 3.x doesn't compile with node12, which Meteor 1.9 uses.

Meteor did not automatically upgrade us to 4.x when it came out many years ago because it believes major version bumps are backwards-incompatible.

AFAICT, fourseven:scss's version number matches the underlying node-sass version number. node-sass 4.0.0's big change was upgrading to libsass 3.4 which corresponds to Ruby Sass 3.4. It looks to me like it is intended to be backwards-compatible and the major version bump was gratuitous, but I guess we'll see...