pipe_wrap API changed in Node 12
This is an internal node API and we have no business using it! So it's 100% our fault that this broke.

However, last I checked, Node has no public API for receiving SCM_RIGHTS file descriptors. We could add our own (probably to node-capnp), but that's a lot of work. Using this internal API and dealing with occasional breakages is still easiest.