Add a tsconfig.json, make typescript happy.
Apparently the meteor typescript plugin doesn't actually do any
*checking* with typescript -- all it does is translate .ts files
into .js files so that the build works.

This patch adds a tsconfig.json and has the `lint` task run tsc
as well (with noEmit: true, so we're just doing the checking, letting
meteor take care of the translation).

The config is borrowed from the output of `meteor create --typescript`.

After adding this and running tsc, typescript complained about two
things:

1. `Crytpo = Npm.require('crypto')` in various parts of the server,
   overwriting the built-in `Crypto` global that is available in the
   browser.
2. The unused `Template` import in `client/main.ts`.

Unfortunately, the config format's "lib" option doesn't allow us to
specify different values for different subdirectories, so it thinks
the DOM apis are available everywhere. To solve (1) we just replaced
these statements with a proper import:

    import Crypto from 'crypto';

To solve (2), we add the Template import to the modules that actually
use it, and remove the variable from `client/main.ts`. We keep the
import of that module in place, so as not to disturb import order,
though I'm not sure it matters in this case, maybe we can remove it
entirely.