Fix flakiness from race condition in installApp test helper
When navigating to /apps, it's possible that the introjs overlay will make the
.app-list element not "visible" (because the overlay is blocking the entirety
of the page).

It's also possible that we'll win that race, but then once we navigate to the
grain list page, we'll lose a race against a different introjs overlay.

Additionally, the introjs overlay does not appear to get reactively removed
when we mark that it's no longer needed in the Session.  So we need to catch
things early.

Making the call to `disableGuidedTour()` *before* navigating to any of the
pages that feature introjs helpers should help ensure that they never render
and never cause test failures.

It's possible we should rethink how we manage introjs and testing -- perhaps we
should disable the overlays earlier for testing purpose, like in user login.
There's only two tests that actually bother interacting with the introjs
overlays, and perhaps they could use a different login method (or pass a
parameter to decline disabling all introjs overlays).

I believe this patch may fix the flakiness around `apps/collections` that we're
seeing in the checks for #3293.