tests/apps/schedule.js: wait a bit before calling runDueJobsAt.
@zarvox pointed out that there's a race condition where this happens
before the job is actually inserted into the database.

Strictly speaking, inserting a .pause here doesn't actually fix the race
condition, but (a) the test is using this technique in several other
places, and (b) If it takes 5 seconds for the job to hit the DB
something strange is going on.

It would be nice to re-work the test to avoid these pauses, if only so
that it will run faster, but if the race condition is the cause of the
issues we're seeing, this should at least get things passing again.