Add more packages.
    meteor npm install --save "@root/csr" "@root/pem" "@root/keypairs" \
        acme-dns-01-cloudflare acme-dns-01-digitalocean acme-dns-01-dnsimple \
        acme-dns-01-duckdns acme-dns-01-godaddy acme-dns-01-gandi \
        acme-dns-01-namecheap acme-dns-01-namedotcom acme-dns-01-route53 \
        acme-dns-01-vultr

The first three packages are from the same author as ACME.js and intended to be used along with it.

The rest are plugins for ACME.js to handle various DNS providers.

It turns out that all of these plugins are also by the authors of ACME.js *except for* the Cloudflare plugin, which is by a Cloudflare power user that I've interacted with before. (If each one were by a different rando then I'd be less comfortable adding all these...)

I will design the code such that an ACME plugin won't be loaded at all unless the server admin specifically chooses to use it.