Fix bitrot in install docs.
Per comments in the Makefile and the GitHub action, we actually need to
install specifically meteor 1.8.2, to work around a problem with font
generation (ick). The installing from source docs were never updated
when the workaround became necessary however; this patch does that.

See also #3303, where someone tripped on this getting set up to hack on
things.