Docs: Provide advice for AD-over-LDAP config
The default searchUsername populated by Sandstorm is `uid`, which is not populated in Active Directory. I figured we should add an implementation note here on what setting(s) someone should consider when trying to use LDAP with Active Directory. I believe both choices presented in this PR are potentially valid choices depending on the environment.