Fix broken use of Settings.
This should fix #3322, but I don't have an LDAP setup to test with.

This was caused by the recent refactoring -- we added an explicit import
for `Settings`, but no explicit *export*.

This is actually the only remaining use of the variable, so instead of
simply revert the broken refactor, I just switched it over to using
`globalDb.collection.settings` (which is used all over the place), and
removed the `Settings` alias entirely.