Port some of the identicon stuff over to typescript.
Just identicon.js for now, the rest will come in future commits.

This was mostly just a matter of adding annotations in the right places,
but I also had to rename a local variable s (to `sat`), because it was
being re-assigned to a value of a different type (number to number[]).
There should be no functional impact.

The identicon module is interesting mostly because it is the only
dependency of sandstorm-db/db.js. The db stuff will require more
refactoring to get to a place where typescript will be happy with it,
but it will also have much bigger payoffs.