tests: apps/schedule: avoid a call to .pause()
...by implementing the proper fix for the race condition described in

As expected, this shaves about 5 seconds off of each of these tests, or
about 15 seconds off of the test suite as a whole.

Note that there are two other calls to pause() in there; these are a bit
harder to eliminate, since the events they're waiting for are triggered
by things outside of any http request, so we can't easily create an
element on the page and do waitForVisible or the like, and as far as I
can tell nightwatch doesn't give us an easy way to say "wait until this
element contains this text," which is what we really want here.