Silence a bunch of unused-argument warnings
A migration is just a function that takes a SandstormDb and a backend,
but the majority of these functions don't actually use the backend.

This patch updates our eslint config to allow silencing an unused
argument warning by prefixing the argument with an underscore, then uses
this throughout migrations.js. This way we can start cutting down on
the noise.