Allow scrolling the navbar
When there are many grains open on desktop or just a few grains open on Firefox on Android, the user needs to be able to scroll the navbar menu down to see them. These changes should allow that without breaking anything else. It was a little difficult to trace all the curly brackets for scss, but here are the intended changes:

@media (max-width: 900px) {
    body > .topbar {
        overflow: hidden; /* REMOVE THIS */
        overflow: scroll; /* ADD THIS */
    }
}
@media (min-width: 901px) {
     body > .topbar > .navbar {
        overflow: scroll; /* ADD THIS */
    }
}