Updated cron script for extracting certificates
The original script to extract Sandcats certificates no longer works,
per Sandstorm issue #3370.  This is an attempt at a new script that
appears to work properly.

The new script could also be copied to a repository owned by the
Sandstorm project or one of the committers, and it's trivial shell
commands so I'm happy to relicense it upon request.