Disable HTTP access through hack-session.
Per the comments in db.js, it is still possible to re-enable this for
now, through an un-documented database setting that users will have to
ask us about to learn of. In the comments for #3192 we'd planned on
making this an option in sandstorm.conf, but this is a bit easier to
implement and given that we're intentionally avoiding documenting this,
requiring people who need it to pop open `sandstorm mongo` seems fine.

When we finally remove this, we'll want to add a database migration that
deletes the setting.