Fix a path traversal vulnerability involving static publishing.
Previously, getWwwFileHack() would unconditionally follow symbolic
links.  An attacker able to place symlinks in a grains /var/www could
abuse this to perform a path traversal attack, allowing regular files
visible to the supervisor but outside of the static publishing directory
to be accessed via static publishing.

This is limited to the grain's mutable storage, since the supervisor
self-sandboxes, but notably it does include the parent directory of the
sandbox, which is not normally accessible to the grain. This directory
contains the grain log.

The damage is mostly confined to the grain itself, as the supervisor has
access to little else besides the grain's storage, but this could still
be used to extract confidential information from a grain that allows
less-privileged users of the grain to upload files for static
publishing, if the upload mechanism allows symlinks.

With this patch, we avoid following symlinks automatically and instead
resolve them manually, refusing to follow any link that points outside
of www.