Correctly handle responses to HEAD requests that use Content-Length
Previously, if the bridge saw a Content-Length header, but the length of
the body did not match, it would throw an error. However, for HEAD
requests this is incorrect, so with this patch the check only applies to
non-HEAD requests; for HEAD requests we will expect an empty body.

This also changes WebSessionBridge to ignore the body of HEAD requests

Fixes #3454