Generate JavaScript source maps.
This will give us accurate stack traces in the browser, and generally
make debugging client-side code easier.

Note that there is a pr to merge this functionality into meteor core:

https://github.com/meteor/meteor/pull/10341

...but which is stalled on some things Sandstorm doesn't actually care
about (notably, many developers don't actually want to expose the source
to end-users, they want the source maps uploaded to a CI or something,
but Sandstorm is FOSS so we absolutely _do_ want users to have the
source).

The PR has been open for a couple years though. If/when that lands we
can switch over to using meteor core's support.