Fix #3484
When oidc support was added, we never updated the setup wizard to supply
this now-mandatory parameter.

I tested this manually and confirmed that it fixes the problem, but
soonish I want to write some tests for the setup wizard in particular,
as this isn't the first time we've introduced bugs that only affect
setup.