Improve handling of ioctl().
This allows a couple things we didn't before, and adjust error returns
for some others.

In particular, the close-on-exec ioctls are used by libuv, which is used
by node, which is probably why Rocket.Chat wasn't starting previously --
it gets a bit farther now, but it looks like mongo is having issues now,
which I will need to investigate.