Fix a typo in a variable name.
I spotted this when it threw an exception while I was doing other
development. So, this fixes a bug where the intended logic was to stop
re-trying after some number of consecutive disconnects, but because of
the typo, tasks would stay in the "retry queue" forever, unless they
were eventually successful.

Not a super serious bug, since retries don't happen rapidly anyway,
but...