Fix use-after-free caught by deleting StringPtr::operator=(String&&).
The operator was deleted in a recent change to Cap'n Proto, to detect bugs like this one.

In this particular case, I do not think any data leak or corruption of pointers was possible, because immediately after the string buffer was deleted, it was read in order to create a copy of the string. The memory couldn't possibly have been reused for anything else in that time. The malloc implementation itself might have overwritten the buffer with a pointer or something, but not any kind of sensitive secret. The UAF was read-only.

It looks like the worst case is this caused corruption of filenames when moving files via the DAV protocol. But I suspect that in practice it caused no actual issues.