Vagrantfile: Create a private network only on virtualbox
The block in question gets evaluated regardless of virtualization provider,
and prior to this change, its presence would cause the following error
under libvirt:

    Network 192.254.254.2 is not available. Specify available network
    name, or an ip address if you want to create a new network.

This change adjusts it so the network is only created under virtualbox.

Note that, under libvirt, this private network does not even seem to be
necessary to achieve an NFS share.  Excerpt from the output on such a machine:

    ==> default: Installing NFS client...
    ==> default: Exporting NFS shared folders...
    ==> default: Preparing to edit /etc/exports. Administrator privileges will be required...
    [sudo] password for garrison:
    ==> default: Mounting NFS shared folders...