Make the grain close button bigger on mobile.
...Filling the entire vertical space alotted for it.

I have found this difficult to press at its current size on mobile; this
essentially doubles the size, so it should be better.

Aesthetically it would be nice to also enlarge the grain icon for
symmetry, but just applying the same rule ends up with it overflowing
the bounds significantly. I am trying to resist the temptation to redo
this whole section with flexbox & grid.